meals interactors impl

+++ domain/src/main/java/com/crskdev/mealcalculator/domain/entities/Meal.kt
@@ -9,5 +9,5 @@ data class Meal(val id: Int = -1,
                 val carbohydrate: Carbohydrate,
                 val fat: Fat,
                 val protein: Float,
-                val glycemicLoad: Int,
+                val glycemicLoad: Float,
                 val date: String? = null)
